Check if the correct class is submitted by using it
Probably the best way to check if we can use the class of the student in
the test, is using it in the tests. This commit adds a simple test which
does just that. There's only a single reason this test could fail, which
simplifies reporting on it. If importing the class in this test works,
we can assume it will in future tests.