Use a SecurityManager and ExitException to handle System.exit-students
When using System.exit(...), an ExitException will be thrown. This has
the same effect from the user's point of view (the code exits) as it
winds up call hierarchy to crash, with the exception it can be caught
and checked by the test code. If the ExitException is not caught, the
submission is evaluated as incorrect: students shouldn't use System.exit
in their code unless asked.